requirements.txt
black
flake8
ipython
isort
jupyterlab
loguru
matplotlib
seaborn
notebook
numpy
pandas
pip
python-dotenv
tqdm
datargs
torch
torchvision
lightning
pillow
typer
wandb
torchmetrics
-e .
